Tiles Save from Editor (no items or nests yet)
When editting maps, be sure to be working with the game launched from
visual studio. The levels save into the content folder within the
project and not the current build. If you edit from the current build,
when the project is built again then the levels in the project will
override those in the current build.